requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.1
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
chromadb==0.5.23
unstructured
sentence-transformers
python-dotenv
streamlit
